+++ sdk/include/vcruntime/_mingw_mac.h
