+++ sdk/include/vcruntime/comdef.h
