+++ sdk/include/vcruntime/comdefsp.h
