+++ sdk/include/vcruntime/comutil.h
