+++ sdk/include/vcruntime/crtdefs.h
