+++ sdk/include/vcruntime/gcc_sal.h
