+++ sdk/include/vcruntime/mingw32/gcc_float.h
