+++ sdk/include/vcruntime/mingw32/intrin_ppc.h
