+++ sdk/include/vcruntime/ms_sal.h
