+++ sdk/include/vcruntime/msc/msc_float.h
