+++ sdk/include/vcruntime/rtcapi.h
