+++ sdk/include/vcruntime/sal.h
