+++ sdk/include/vcruntime/stdbool.h
