+++ sdk/include/vcruntime/vadefs.h
