+++ sdk/include/vcruntime/varargs.h
